+++ src/dynamic_rrt/src/rrt_cuda.hpp
@@ -13,7 +13,7 @@ public:
     RRTCuda();
 
 private:
-    auto set_resolution(double resolution) -> void override;
+    auto set_resolution(float resolution) -> void override;
 
     auto plan_rrt(
         const Pose2D& start,
